adding support for floats

+++ core/vm/arch/jit/arm64/jit_arm_a64.h
@@ -252,7 +252,7 @@ namespace Runtime {
       return type;
     }
 
-    int32_t GetOperand() {
+    long GetOperand() {
       return operand;
     }
   };
